Split sensors from suite_carla.py to carla_sensors.py
Because suite_carla.py is getting longer.
